+++ doc/whats_new.rst
@@ -54,6 +54,8 @@ Changelog
 Bug
 ~~~
 
+- Fix :meth:`mne.io.Raw.append` annotations miss-alignment  by `Joan Massich`_
+
 - Fix :func:`mne.io.read_raw_edf` reading duplicate channel names by `Larry Eisenman`_
 
 - Fix :func:`set_bipolar_reference` in the case of generating all bipolar combinations and also in the case of repeated channels in both lists (anode and cathode) by `Cristóbal Moënne-Loccoz`_